Fix: Slow page refresh when an MV200 is offline
The MV200 list query API calls the DPUAgent of all MV200s to obtain information
about whether cloud disk boot is supported. However, if an MV200 is offline,
this increases the overall API response time. Multiple pages also call this API,
causing slow page display.

The backend API for querying the MV200 list no longer calls the DPUAgent of all
MV200s to obtain information about whether cloud disk boot is supported. This
information is now queried individually for each MV200 on the page that needs it.
This ensures that the list query is not affected by whether an MV200 is offline or not.