Added fallback
FRU information was originally only used to identify the board type via the 90 part number,
but if the SN is not in lspci, it is retrieved from the FRU information.

Since both of these pieces of information are in the FRU, why not just use FRU directly instead
of retrieving them from lspci first? Because the yuncli command might be unavailable, lspci is
used as a fallback.